# Recoverability doesn't remove all error message headers when retrying