--- a/src/ServiceControl/Recoverability/Retrying/HeaderFilter.cs
+++ b/src/ServiceControl/Recoverability/Retrying/HeaderFilter.cs
@@ -1,16 +1,25 @@
 ﻿namespace ServiceControl.Recoverability
 {
+    using System;
     using System.Collections.Generic;
-    using System.Linq;
 
     static class HeaderFilter
     {
         public static Dictionary<string, string> RemoveErrorMessageHeaders(Dictionary<string, string> headers)
         {
-            var headersToRetryWith = headers
-                .Where(kv => !KeysToRemoveWhenRetryingAMessage.Contains(kv.Key))
-                .ToDictionary(kv => kv.Key, kv => kv.Value);
-
+            // still take a copy to preserve the old assumptions
+            var headersToRetryWith = new Dictionary<string, string>(headers);
+            foreach (var headerToRemove in KeysToRemoveWhenRetryingAMessage)
+            {
+                // iterate over original so that we are not running into modified collection problem
+                foreach (var keyValuePair in headers)
+                {
+                    if (keyValuePair.Key.StartsWith(headerToRemove, StringComparison.Ordinal))
+                    {
+                        headersToRetryWith.Remove(keyValuePair.Key);
+                    }
+                }
+            }
             return headersToRetryWith;
         }
 
@@ -19,21 +28,9 @@
             "NServiceBus.Retries",
             "NServiceBus.FailedQ",
             "NServiceBus.TimeOfFailure",
-            "NServiceBus.ExceptionInfo.ExceptionType",
-            "NServiceBus.ExceptionInfo.AuditMessage",
-            "NServiceBus.ExceptionInfo.Source",
-            "NServiceBus.ExceptionInfo.StackTrace",
-            "NServiceBus.ExceptionInfo.HelpLink",
-            "NServiceBus.ExceptionInfo.Message",
-            "ServiceControl.EditOf",
-            "NServiceBus.ExceptionInfo.Data.Message type",
-            "NServiceBus.ExceptionInfo.Data.Handler type",
-            "NServiceBus.ExceptionInfo.Data.Handler start time",
-            "NServiceBus.ExceptionInfo.Data.Handler failure time",
-            "NServiceBus.ExceptionInfo.Data.Message ID",
-            "NServiceBus.ExceptionInfo.Data.Transport message ID",
-            "NServiceBus.Retries.Timestamp",
-            "NServiceBus.Timeout.Expire"
+            "NServiceBus.ExceptionInfo.",
+            "NServiceBus.Timeout.Expire",
+            "ServiceControl.EditOf"
         };
     }
 }