--- a/src/ServiceControl/Recoverability/Retrying/HeaderFilter.cs
+++ b/src/ServiceControl/Recoverability/Retrying/HeaderFilter.cs
@@ -29,7 +29,6 @@
             "NServiceBus.FailedQ",
             "NServiceBus.TimeOfFailure",
             "NServiceBus.ExceptionInfo.",
-            "NServiceBus.Timeout.Expire",
             "ServiceControl.EditOf"
         };
     }