Added exception, retries and timeout headers (simplest fix)

https://github.com/Particular/ServiceControl/issues/2259