@danielmarbach 2020-12-04
I remove the timeout header for now since I'm not 100% sure we should remove it. I pinged the relevant internal channels