@danielmarbach 2020-12-04
I thought about the string match too but considered it too dangerous because it could remove things that we don't want to have removed. Given that previously the linq did

```
var headersToRetryWith = headers
                .Where(kv => !KeysToRemoveWhenRetryingAMessage.Contains(kv.Key))
                .ToDictionary(kv => kv.Key, kv => kv.Value);
```

I think we are in similar big O territory as the current solution so I'm not too worried. Thoughts?