@danielmarbach 2020-12-04
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 9 3950X, 1 CPU, 32 logical and 16 physical cores
.NET Core SDK=3.1.403
  [Host]     : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  DefaultJob : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT


```
| Method |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |      Op/s | Ratio |
|------- |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|----------:|------:|
| Before | 5.133 μs | 0.0358 μs | 0.0335 μs | 0.0087 μs | 5.086 μs | 5.098 μs | 5.136 μs | 5.160 μs | 5.189 μs | 194,801.4 |  1.00 |
|  After | 2.693 μs | 0.0446 μs | 0.0418 μs | 0.0108 μs | 2.636 μs | 2.670 μs | 2.677 μs | 2.706 μs | 2.785 μs | 371,375.6 |  0.52 |

I think that says it all