@timbussmann 2020-12-04
> I think we are in similar big O territory as the current solution so I'm not too worried. Thoughts?

yeah I'm totally ok with the proposed solution 👍